Default to https for URLs without protocols in parseLinksFromText
This causes mixed content errors (e.g. ERR_CLEARTEXT_NOT_PERMITTED) when opening auto-linkified links in messages where I have defined a custom link component in the markdown rules so that links open within the app using a web view.

Since the web is increasingly moving to `https` as the default over `http`, I think it would be better to use `https` over `http` when defining a protocol for a URL that doesn't have one in the message.